min/max float->double
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@10583 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ m_config.h
@@ -71,7 +71,7 @@ struct config {
   void *p; 
   struct m_option_type* type;
   unsigned int flags;
-  float min,max;
+  double min,max;
   void* priv;
 };
